Sort by all coordinates at the same time.
Instead of sorting by one coordinate, segmenting, and sorting by the
next coordinate, sort by all three coordinates simultaneously.
Unfortunately, this is just a bit slower though not noticeably so
(Δ=+0.3s over 10s total for 12M points). The upside is we no longer need
two extra working arrays. Note, this does produce a slightly different
order for elements that are the same, but since heap sort is not stable,
that's to be expected.